Since we allocate an extra byte (for terminating NUL), overflow protection has to keep the needed string length strictly less than INT_MAX.
EDIT: Sorry, that was wrong, for a slightly tricky reason. The objPtr->length
field is type int, so must be limited to INT_MAX. That's the limit to protect.
The memory allocation routines accept (unsigned int) argument as the number
of bytes to allocate, so they can accept a request for INT_MAX + 1.